Here's about 1/3 of all the warnings from cmake squashed. WARNING: There were/are many BUGS which I could not properly fix; namely variables that were declared without initialization, then initialize inside some conditionals or cases but not others, and subsequently used. In 90% of the cases I initialized them to zero, or to 0.0f, or to enum Ok, etceteras; but this NEEDS attention from someone more familiar with the code. Check the diffs. Many files like that. Warnings from the collision library were made to shut up with a pragma. They rely on sizeof(float)==sizeof(int) and (int&)float(0.0f)==int(0), which happens to be true, currently, with AMD athlons and compatible cpu's. Terribly incorrect library, tho.

+++ src/gfx/background.cpp
@@ -329,7 +329,7 @@ void Background::Draw()
             skybox_rendering_sequence[3].tex = right;
             skybox_rendering_sequence[4].tex = back;
             skybox_rendering_sequence[5].tex = down;
-            for (int skr = 0; skr < sizeof (skybox_rendering_sequence)/sizeof (skybox_rendering_sequence[0]); skr++) {
+            for (size_t skr = 0; skr < sizeof (skybox_rendering_sequence)/sizeof (skybox_rendering_sequence[0]); skr++) {
                 Texture *tex = skybox_rendering_sequence[skr].tex;
                 int lyr;
 
@@ -384,19 +384,19 @@ void Background::Draw()
 
 #ifdef NV_CUBE_MAP
               #define VERTEX( i )                                    \
-    s1 = stca[skybox_rendering_sequence[skr].tcoord[i][0]];          \
-    t1 = ttca[skybox_rendering_sequence[skr].tcoord[i][1]];          \
-    s2 = stca[skybox_rendering_sequence[skr].tcoord[i][2]];          \
+    s1 = stca[size_t(skybox_rendering_sequence[skr].tcoord[i][0])];          \
+    t1 = ttca[size_t(skybox_rendering_sequence[skr].tcoord[i][1])];          \
+    s2 = stca[size_t(skybox_rendering_sequence[skr].tcoord[i][2])];          \
     glTexCoord3f( s1, t1, s2 );                                      \
     GFXVertex3f( skybox_rendering_sequence[skr].vertices[i][0]*size, \
                  skybox_rendering_sequence[skr].vertices[i][1]*size, \
                  skybox_rendering_sequence[skr].vertices[i][2]*size );
 #else
               #define VERTEX( i )                                    \
-    s1 = stca[skybox_rendering_sequence[skr].tcoord[i][0]];          \
-    t1 = ttca[skybox_rendering_sequence[skr].tcoord[i][1]];          \
-    s2 = stca[skybox_rendering_sequence[skr].tcoord[i][2]];          \
-    t2 = ttca[skybox_rendering_sequence[skr].tcoord[i][3]];          \
+    s1 = stca[size_t(skybox_rendering_sequence[skr].tcoord[i][0])];          \
+    t1 = ttca[size_t(skybox_rendering_sequence[skr].tcoord[i][1])];          \
+    s2 = stca[size_t(skybox_rendering_sequence[skr].tcoord[i][2])];          \
+    t2 = ttca[size_t(skybox_rendering_sequence[skr].tcoord[i][3])];          \
     if (!multitex) GFXTexCoord2f( s1, t1 );                          \
     else GFXTexCoord4f( s1, t1, s2, t2 );                            \
     GFXVertex3f( skybox_rendering_sequence[skr].vertices[i][0]*size, \
